etl/requirements.txt
duckdb>=0.10.0
dbfread>=2.0.7
osmnx>=1.9.0,<3.0.0
psycopg2-binary>=2.9.0
